requirements.txt
google-adk==1.9.0
litellm==1.75.0
